Add Retrofit for interaction with REST APIs on the server
JDK version updated to JDK 8 for Retrofit to work properly. A LoginAPI along with
corresponding payload has been defined. RetrofitAPIBuilder can be used to construct a
defined API for calls to the server. LoginActivity has code added to send a POST request
to the /login endpoint when FCM is enabled.