Add login and join session calls to server
Login call is added so that the server can verify that the user is a valid Facebook
user and the Firebase token provides a way to push data to the client.
New dialog is added in MainActivity for users to input their session ID. A call to the
server is made to join the given session before moving to SessionActivity.