Update *Repository and add ListViewModel
CommonRunnables provides common Runnables that can be provided as arguments for calls to Repository.
ListRepository is now more fleshed out with the ability to get list/lists.
ListViewModel is used as a layer between Activities and Repositories.
Added new SessionApi call to update the list for a session.
Fix compile error for SwipeActivity now that a session does not have a list of ideas.