Add FirebaseIntentReceiver for ListPayload to SessionActivity
Since the server can sometimes return messages that contain ListPayload,
and the lists are necessary for starting a session, a IntentReceiver
is required to receive the Intents sent by the Service.

Note that the order of broadcasts sent by the Service matters: we need to
ensure that the list in the repository is populated before we see that the session
has been started, otherwise we may start a session with an empty list of ideas.