App retrieves list prior to opening SwipeActivity
No longer rely on list being sent from back-end as part of Firebase push since Firebase messages
are limited in size, so long lists could cause the message to be rejected.
Instead we retrieve the list after we see that the session has been started, so that it is
populated for when we start swiping. As a result, MyFirebaseMessagingService no longer
needs to handle receiving a list and we no longer need to have Receivers for lists.