Implement deleting and persistence of entries
Delete button on each RecyclerView item now properly deletes from list of entries.
ViewHolder takes in an IdeaPayload when binded and displays it, then updates the IdeaPayload
after it is recycled so that whatever the user inputs is stored. Since the ViewHolder knows
which position it is, upon press of the delete button, it can let the adapter know which
list index to delete.