Implement view and editing of specific lists
Users can now click on the edit button to enter a new activity for editing a list.
Refactor CreateNewListActivity to handle editing of existing lists as well.
The activity checks for a flag in the intent to figure out what it should be displaying.
Classes and variable names have been renamed to be a bit clearer (ex. ViewEditListsActivity to
ViewOrUpdateListsActivity)